+++ CHANGES.rst
@@ -10,6 +10,8 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 4.8.6 (unreleased)
 ------------------
 
+- Adjust the old ``App.Dialogs.MessageDialog`` to be served as HTML after
+  from PR https://github.com/zopefoundation/Zope/pull/1075 .
 
 4.8.5 (2022-12-17)
 ------------------